Fixed C compiler warning for convertion between char and int.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@38287 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/eif_type_id.c
@@ -93,7 +93,7 @@ rt_public EIF_TYPE_ID eif_type_id (char *type_string)
 
 rt_private int is_good (char c)
 {
-	return isalpha (c) || isdigit (c) || (c == '_');
+	return isalpha ((int) c) || isdigit ((int) c) || (c == '_');
 }
 
 rt_private int is_type_separator (char c)
